pyproject.toml
[tool.poetry]
name = "stackdiac"
version = "0.0.1-dev6"
description = ""
authors = ["sysr9 <38893296+sysr9@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
coloredlogs = "^15.0.1"
pyyaml = "^6.0"
pydantic = "^1.10.6"
gitpython = "^3.1.31"
requests = "^2.28.2"
jinja2 = "^3.1.2"
deepmerge = "^1.1.0"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
mergedeep = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
stackd = "stackdiac.cli:cli"